+++ pkg/cmd/cmd.go
@@ -29,7 +29,7 @@ import (
 	"github.com/vmware-tanzu/cartographer/pkg/controller/delivery"
 	"github.com/vmware-tanzu/cartographer/pkg/controller/runnable"
 	"github.com/vmware-tanzu/cartographer/pkg/controller/supplychain"
-	"github.com/vmware-tanzu/cartographer/pkg/controller/workload"
+	"github.com/vmware-tanzu/cartographer/pkg/controllers"
 	"github.com/vmware-tanzu/cartographer/pkg/utils"
 )
 
@@ -83,7 +83,7 @@ func (cmd *Command) Execute(ctx context.Context) error {
 }
 
 func registerControllers(mgr manager.Manager) error {
-	if err := (&workload.Reconciler{}).SetupWithManager(mgr); err != nil {
+	if err := (&controllers.Reconciler{}).SetupWithManager(mgr); err != nil {
 		return fmt.Errorf("failed to register workload controller: %w", err)
 	}
